+++ src/modules/graphics/opengl.c
@@ -1,4 +1,3 @@
-#include "graphics/opengl.h"
 #include "graphics/graphics.h"
 #include "graphics/buffer.h"
 #include "graphics/canvas.h"
@@ -16,6 +15,15 @@
 #include <string.h>
 #include <stdlib.h>
 #include <stdio.h>
+
+#ifdef LOVR_WEBGL
+#include <GLES3/gl3.h>
+#include <GLES2/gl2ext.h>
+#include <GL/gl.h>
+#include <GL/glext.h>
+#else
+#include "lib/glad/glad.h"
+#endif
 
 // Types
 
@@ -97,6 +105,28 @@ struct Shader {
   map_t uniformMap;
   map_t blockMap;
   bool multiview;
+};
+
+struct Mesh {
+  uint32_t vao;
+  uint32_t ibo;
+  DrawMode mode;
+  char attributeNames[MAX_ATTRIBUTES][MAX_ATTRIBUTE_NAME_LENGTH];
+  MeshAttribute attributes[MAX_ATTRIBUTES];
+  uint8_t locations[MAX_ATTRIBUTES];
+  uint16_t enabledLocations;
+  uint16_t divisors[MAX_ATTRIBUTES];
+  map_t attributeMap;
+  uint32_t attributeCount;
+  struct Buffer* vertexBuffer;
+  struct Buffer* indexBuffer;
+  uint32_t vertexCount;
+  uint32_t indexCount;
+  size_t indexSize;
+  size_t indexOffset;
+  uint32_t drawStart;
+  uint32_t drawCount;
+  struct Material* material;
 };
 
 typedef enum {
@@ -2797,7 +2827,8 @@ Buffer* lovrShaderBlockGetBuffer(ShaderBlock* block) {
 
 // Mesh
 
-Mesh* lovrMeshInit(Mesh* mesh, DrawMode mode, Buffer* vertexBuffer, uint32_t vertexCount) {
+Mesh* lovrMeshCreate(DrawMode mode, Buffer* vertexBuffer, uint32_t vertexCount) {
+  Mesh* mesh = lovrAlloc(Mesh);
   mesh->mode = mode;
   mesh->vertexBuffer = vertexBuffer;
   mesh->vertexCount = vertexCount;
@@ -2831,4 +2862,124 @@ void lovrMeshSetIndexBuffer(Mesh* mesh, Buffer* buffer, uint32_t indexCount, siz
     mesh->indexSize = indexSize;
     mesh->indexOffset = offset;
   }
+}
+
+Buffer* lovrMeshGetVertexBuffer(Mesh* mesh) {
+  return mesh->vertexBuffer;
+}
+
+Buffer* lovrMeshGetIndexBuffer(Mesh* mesh) {
+  return mesh->indexBuffer;
+}
+
+uint32_t lovrMeshGetVertexCount(Mesh* mesh) {
+  return mesh->vertexCount;
+}
+
+uint32_t lovrMeshGetIndexCount(Mesh* mesh) {
+  return mesh->indexCount;
+}
+
+size_t lovrMeshGetIndexSize(Mesh* mesh) {
+  return mesh->indexSize;
+}
+
+uint32_t lovrMeshGetAttributeCount(Mesh* mesh) {
+  return mesh->attributeCount;
+}
+
+void lovrMeshAttachAttribute(Mesh* mesh, const char* name, MeshAttribute* attribute) {
+  uint64_t hash = hash64(name, strlen(name));
+  lovrAssert(map_get(&mesh->attributeMap, hash) == MAP_NIL, "Mesh already has an attribute named '%s'", name);
+  lovrAssert(mesh->attributeCount < MAX_ATTRIBUTES, "Mesh already has the max number of attributes (%d)", MAX_ATTRIBUTES);
+  lovrAssert(strlen(name) < MAX_ATTRIBUTE_NAME_LENGTH, "Mesh attribute name '%s' is too long (max is %d)", name, MAX_ATTRIBUTE_NAME_LENGTH);
+  lovrGraphicsFlushMesh(mesh);
+  uint64_t index = mesh->attributeCount++;
+  mesh->attributes[index] = *attribute;
+  strcpy(mesh->attributeNames[index], name);
+  map_set(&mesh->attributeMap, hash, index);
+  lovrRetain(attribute->buffer);
+}
+
+void lovrMeshDetachAttribute(Mesh* mesh, const char* name) {
+  uint64_t hash = hash64(name, strlen(name));
+  uint64_t index = map_get(&mesh->attributeMap, hash);
+  lovrAssert(index != MAP_NIL, "No attached attribute named '%s' was found", name);
+  MeshAttribute* attribute = &mesh->attributes[index];
+  lovrGraphicsFlushMesh(mesh);
+  lovrRelease(Buffer, attribute->buffer);
+  map_remove(&mesh->attributeMap, hash);
+  mesh->attributeNames[index][0] = '\0';
+  memmove(mesh->attributeNames + index, mesh->attributeNames + index + 1, (mesh->attributeCount - index - 1) * MAX_ATTRIBUTE_NAME_LENGTH * sizeof(char));
+  memmove(mesh->attributes + index, mesh->attributes + index + 1, (mesh->attributeCount - index - 1) * sizeof(MeshAttribute));
+  mesh->attributeCount--;
+  for (uint32_t i = 0; i < MAX_ATTRIBUTES; i++) {
+    if (mesh->locations[i] > index) {
+      mesh->locations[i]--;
+    } else if (mesh->locations[i] == index) {
+      mesh->locations[i] = 0xff;
+    }
+  }
+}
+
+const MeshAttribute* lovrMeshGetAttribute(Mesh* mesh, uint32_t index) {
+  return index < mesh->attributeCount ? &mesh->attributes[index] : NULL;
+}
+
+uint32_t lovrMeshGetAttributeIndex(Mesh* mesh, const char* name) {
+  uint64_t hash = hash64(name, strlen(name));
+  uint64_t index = map_get(&mesh->attributeMap, hash);
+  return index == MAP_NIL ? ~0u : index;
+}
+
+const char* lovrMeshGetAttributeName(Mesh* mesh, uint32_t index) {
+  return mesh->attributeNames[index];
+}
+
+bool lovrMeshIsAttributeEnabled(Mesh* mesh, const char* name) {
+  uint64_t hash = hash64(name, strlen(name));
+  uint64_t index = map_get(&mesh->attributeMap, hash);
+  lovrAssert(index != MAP_NIL, "Mesh does not have an attribute named '%s'", name);
+  return !mesh->attributes[index].disabled;
+}
+
+void lovrMeshSetAttributeEnabled(Mesh* mesh, const char* name, bool enable) {
+  bool disable = !enable;
+  uint64_t hash = hash64(name, strlen(name));
+  uint64_t index = map_get(&mesh->attributeMap, hash);
+  lovrAssert(index != MAP_NIL, "Mesh does not have an attribute named '%s'", name);
+  if (mesh->attributes[index].disabled != disable) {
+    lovrGraphicsFlushMesh(mesh);
+    mesh->attributes[index].disabled = disable;
+  }
+}
+
+DrawMode lovrMeshGetDrawMode(Mesh* mesh) {
+  return mesh->mode;
+}
+
+void lovrMeshSetDrawMode(Mesh* mesh, DrawMode mode) {
+  mesh->mode = mode;
+}
+
+void lovrMeshGetDrawRange(Mesh* mesh, uint32_t* start, uint32_t* count) {
+  *start = mesh->drawStart;
+  *count = mesh->drawCount;
+}
+
+void lovrMeshSetDrawRange(Mesh* mesh, uint32_t start, uint32_t count) {
+  uint32_t limit = mesh->indexSize > 0 ? mesh->indexCount : mesh->vertexCount;
+  lovrAssert(start + count <= limit, "Invalid mesh draw range [%d, %d]", start + 1, start + count + 1);
+  mesh->drawStart = start;
+  mesh->drawCount = count;
+}
+
+Material* lovrMeshGetMaterial(Mesh* mesh) {
+  return mesh->material;
+}
+
+void lovrMeshSetMaterial(Mesh* mesh, Material* material) {
+  lovrRetain(material);
+  lovrRelease(Material, mesh->material);
+  mesh->material = material;
 }
